+++ tests/analyzer/conftest.py
@@ -1,0 +1,1 @@
+from tests.analyzer.fixtures import *
