+++ tests/core_data_structures/test_resampling.py
@@ -15,11 +15,6 @@ _ALL_METHODS = [
 ]
 
 
-# ---------------------------------------------------------------------------
-# _interpolate — boundary clamping (all methods)
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.parametrize("method", _ALL_METHODS)
 def test_interpolate_clamps_before_first(method):
     src_t = np.array([5.0, 10.0])
@@ -36,11 +31,6 @@ def test_interpolate_clamps_after_last(method):
     np.testing.assert_allclose(result, [20.0])
 
 
-# ---------------------------------------------------------------------------
-# _interpolate — exact at knots (LINEAR, NEAREST, CUBIC)
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.parametrize(
     "method",
     [ResampleMethod.LINEAR, ResampleMethod.NEAREST, ResampleMethod.CUBIC],
@@ -50,11 +40,6 @@ def test_interpolate_exact_at_knots(method):
     src_v = src_t**2
     result = _interpolate(src_t, src_t, src_v, method)
     np.testing.assert_allclose(result, src_v, atol=1e-6)
-
-
-# ---------------------------------------------------------------------------
-# _interpolate — LINEAR
-# ---------------------------------------------------------------------------
 
 
 def test_interpolate_linear_interior():
@@ -72,11 +57,6 @@ def test_interpolate_linear_multiple_points():
     np.testing.assert_allclose(result, [0.0, 25.0, 50.0, 75.0, 100.0])
 
 
-# ---------------------------------------------------------------------------
-# _interpolate — ZOH
-# ---------------------------------------------------------------------------
-
-
 def test_interpolate_zoh_holds_previous_value():
     src_t = np.array([0.0, 4.0, 8.0])
     src_v = np.array([10.0, 20.0, 30.0])
@@ -91,22 +71,12 @@ def test_interpolate_zoh_at_knot_returns_knot_value():
     np.testing.assert_allclose(result, [2.0])
 
 
-# ---------------------------------------------------------------------------
-# _interpolate — NEAREST
-# ---------------------------------------------------------------------------
-
-
 def test_interpolate_nearest_close_to_knot():
     src_t = np.array([0.0, 10.0])
     src_v = np.array([0.0, 100.0])
     # 1.0 is closer to 0.0 than to 10.0
     result = _interpolate(np.array([1.0]), src_t, src_v, ResampleMethod.NEAREST)
     np.testing.assert_allclose(result, [0.0])
-
-
-# ---------------------------------------------------------------------------
-# _interpolate — CUBIC
-# ---------------------------------------------------------------------------
 
 
 def test_interpolate_cubic_fallback_fewer_than_4_points():
@@ -121,11 +91,6 @@ def test_interpolate_cubic_interior_smooth():
     src_v = src_t * 2.0
     result = _interpolate(np.array([0.5, 1.5, 2.5]), src_t, src_v, ResampleMethod.CUBIC)
     np.testing.assert_allclose(result, [1.0, 3.0, 5.0], atol=1e-6)
-
-
-# ---------------------------------------------------------------------------
-# resample_to_freq
-# ---------------------------------------------------------------------------
 
 
 @pytest.mark.parametrize(
